requirements.txt
requests
pysocks
websocket-client
